Fixed a bug with the laminar Pr model + a problem in referencing the incorrect cell when actually using the model in Discretization_new.h.
Also, further refinement on the intrusion class. Seems to be working for
multiple patches now, even when geometry lines up with the patch
boundaries.

+++ src/CCA/Components/Arches/BoundaryCondition.cc
@@ -6425,10 +6425,10 @@ BoundaryCondition::velocityOutletPressureBC__NEW( const Patch* patch,
             double negsmall = -1.0E-10;
             double possmall =  1.0E-10;
             double zero     = 0.0E0; 
-            int sign        = 1;
+            double sign        = 1.0;
 
             if ( bc_iter->second.type == PRESSURE ) { 
-              sign = -1; 
+              sign = -1.0; 
             }
 
             switch (face) {
